fix: FolderPicker returns always internal paths on Android🐛
FolderPicker always returned an internal path like '/storage/emulated/0/DCIM' even if an external path was selected, resulting in the app crashing when selecting a folder that only exists on sd card (like /storage/sdCard/External) or selecting the internal path instead of the external (like /storage/sdCard/DCIM resulted to be /storage/emulated/0/DCIM which is the internal one). Now the right path is returned.