Fix Hot Reload with Popup
Call AddLogicalChild/RemoveLogicalChild when the popup is
shown/closed so that popups are included in the logical (and visual)
tree, so that Hot Reload works for them. The Live Visual Tree VS
UI now shows popup content as well.

Fixes https://github.com/CommunityToolkit/Maui/issues/620

As Shane explained to me, the rule as of MAUI in .NET8 is that
whenenever Parent is set on an Element, AddLogicalChild should
also be called to include the element in the logical children.
RemoveLogicalChild should be called when the Parent is set to null.
This may be more automatic in .NET9, but for .NET8 that's what's needed
so that the logical tree is kept up to date.