Replace `fullScreenIcon` with direct `FontIcon` creation
Removed the `fullScreenIcon` static readonly object from the `MauiMediaElement` class in the `MauiMediaElement.windows.cs` file. Now, a new `FontIcon` object is directly created in the `OnFullScreenButtonClick` method. This change simplifies the code and makes the `OnFullScreenButtonClick` method more self-contained.

This removes reduntant static readonly object that was causing issues. The fontIcon is now generated dynamically to prevent CTD.