Add Segoe Fluent Icons font for Windows platform
Updated the `UseMauiCommunityToolkitMediaElement` method in the `AppBuilderExtensions` class to include a configuration for the "Segoe Fluent Icons" font from the "SegMDL2.ttf" file specifically for the Windows platform. Also, added a new `ItemGroup` in the `CommunityToolkit.Maui.MediaElement.csproj` file to pack the "SegMDL2.ttf" font file into the application when the `TargetFramework` contains '-windows'.