Fix Media Element Windows CTD (#1921)
* Updated fullScreenButton icon to fix a bug where App would crash to desktop.

Updated the `Content` property of the `fullScreenButton` object from `fullScreenIcon` to a new `FontIcon` object. The `Glyph` property is now set to "\uE740" and the `FontFamily` property is set to "Segoe Fluent Icons". This change modifies the visual representation of the `fullScreenButton` in the UI.

* Replace `fullScreenIcon` with direct `FontIcon` creation

Removed the `fullScreenIcon` static readonly object from the `MauiMediaElement` class in the `MauiMediaElement.windows.cs` file. Now, a new `FontIcon` object is directly created in the `OnFullScreenButtonClick` method. This change simplifies the code and makes the `OnFullScreenButtonClick` method more self-contained.

This removes reduntant static readonly object that was causing issues. The fontIcon is now generated dynamically to prevent CTD.

* Update full screen button icons in MauiMediaElement

Refactored the `MauiMediaElement` class in `MauiMediaElement.windows.cs` file. Removed the `exitFullScreenIcon` static readonly variable and instead created a new `FontIcon` directly where needed. Updated the `OnFullScreenButtonClick` method to set the `fullScreenButton.Content` to new `FontIcon` instances with different glyphs (`"\uE740"` and `"\uE73F"`), representing different icons in the "Segoe Fluent Icons" font family. These changes affect the icons used for the full screen button in the media player.

This fixes a small UI issue where the incorrect icon was applied starting with full screen then reversion of Full screen status. So icons got reversed starting when selecting full screen. Fixed now. Removed the static readonly object as I was able to verify it does in fact have a bug with it.

* Add Segoe Fluent Icons font for Windows platform

Updated the `UseMauiCommunityToolkitMediaElement` method in the `AppBuilderExtensions` class to include a configuration for the "Segoe Fluent Icons" font from the "SegMDL2.ttf" file specifically for the Windows platform. Also, added a new `ItemGroup` in the `CommunityToolkit.Maui.MediaElement.csproj` file to pack the "SegMDL2.ttf" font file into the application when the `TargetFramework` contains '-windows'.

* Revert adding font directly. Switched back to using existing system font.

* 1. Revert changes
2. remove `static readonly` from `FontIcons`
3. Fix what font icon is applied while switching full screen status

* Add `readonly` to `FontIcons`

* remove uneeded changes from original files