Add destructor to SubtitleExtensions class
Added a destructor method `~SubtitleExtensions()` to the `SubtitleExtensions` class within the `SubtitleExtensions.macios.cs` file. This destructor ensures proper cleanup by unsubscribing from the `FullScreenChanged` event of the `MediaManagerDelegate` and removing a time observer from the `player` object, provided `playerObserver` and `player` are not null. This change aims to prevent potential memory leaks or unwanted behavior by ensuring that instances of `SubtitleExtensions` are correctly disposed of during garbage collection.