Refactor Android context access in Maui app
This commit overhauls how platform-specific activities, windows, and view groups are managed within our Maui application targeting Android. Key changes include the removal of the `PageExtensions` class, which was previously responsible for providing access to the current Android activity, window, and view group. In its place, we've introduced a new `CurrentPlatformContext` record struct within the `MauiMediaElement` class. This struct is designed to serve a similar purpose but with an improved structure and broader access, as indicated by its `public readonly` visibility modifier.

Additionally, the `SubtitleExtensions` class has been updated to utilize the new `CurrentPlatformContext` for managing subtitle functionality, particularly in relation to full-screen mode toggling. This shift away from the now-removed `PageExtensions` class to a more centralized approach within `MauiMediaElement` aims to streamline interactions with the Android platform. It enhances the maintainability and readability of the code, especially for features requiring close integration with Android system capabilities like media playback and full-screen management.