Add AnimateOnTap property to AnimationBehavior
Introduces a new bindable property `AnimateOnTapProperty` to the `AnimationBehavior` class, enabling the addition of a `TapGestureRecognizer` based on user preference. A corresponding public property `AnimateOnTap` is added for getting and setting this value. The `OnAttachedTo` method is updated to conditionally create and add the gesture recognizer. Code formatting and readability improvements are also made by refactoring existing conditions.