Fix CI build errors
Update Label_Loaded method with nullability and logic

The `Label_Loaded` method in `ButtonPopup.xaml.cs` was updated to
accept a nullable `sender` parameter by changing its signature
to `void Label_Loaded(object? sender, EventArgs e)`.

Additionally, the method implementation was added to check if
the `sender` is of type `Label`. If so, it casts the `sender`
to a `Label` and calls the `SetSemanticFocus` method on it.